Enable MySQL TLS
Enable passing the certificate authority on the relation in order to
enable TLS communication to the MySQL DB.

In order to enable MySQL to use TLS for client connections all that is
required is the CA certificate. A certificate and key may be used for
client certificate authentication. However, since almost all deployments
will not use certificate authentication, focusing on the CA as SSL data
complete makes sense.

Change-Id: I1a55e2c342a9f62a6b02da7334e309536bd4eda2